Transferred '_compare_explain_functions' to the 'test_model_explanation.py'
